[SoftwareUpdateChoices] subclass ChoiceBox
