+++ lib/python/Screens/ChoiceBox.py
@@ -6,7 +6,6 @@ from Components.config import config, ConfigSubsection, ConfigText
 from Components.Label import Label
 from Components.ChoiceList import ChoiceEntryComponent, ChoiceList
 from Components.Sources.StaticText import StaticText
-from Components.Pixmap import Pixmap
 import enigma
 from six.moves import zip
 
@@ -17,7 +16,7 @@ config.misc.pluginlist.fc_bookmarks_order = ConfigText(default="")
 
 
 class ChoiceBox(Screen):
-	def __init__(self, session, title="", list=None, keys=None, selection=0, skin_name=None, text="", reorderConfig="", var="", windowTitle=None, allow_cancel=True, titlebartext=_("Choice Box")):
+	def __init__(self, session, title="", list=None, keys=None, selection=0, skin_name=None, text="", reorderConfig="", windowTitle=None, allow_cancel=True, titlebartext=_("Choice Box")):
 		if not windowTitle: #for compatibility
 			windowTitle = titlebartext
 		if not list:
@@ -34,14 +33,6 @@ class ChoiceBox(Screen):
 
 		self.reorderConfig = reorderConfig
 		self["text"] = Label()
-		self.var = ""
-		if skin_name and 'SoftwareUpdateChoices' in skin_name and var and var in ('unstable', 'updating', 'stable', 'unknown'):
-			self.var = var
-			self['feedStatusMSG'] = Label()
-			self['tl_off'] = Pixmap()
-			self['tl_red'] = Pixmap()
-			self['tl_yellow'] = Pixmap()
-			self['tl_green'] = Pixmap()
 
 		title_max = 55
 		if 'MetrixHD/' in config.skin.primary_skin.value:
@@ -147,29 +138,7 @@ class ChoiceBox(Screen):
 		{
 			"back": self.cancel,
 		}, prio=-1)
-		self.onShown.append(self.onshow)
 
-
-	def onshow(self):
-		if self.skinName and 'SoftwareUpdateChoices' in self.skinName and self.var:
-			from Components.OnlineUpdateCheck import feedsstatuscheck
-			if self.var in feedsstatuscheck.feed_status_msgs:
-				status_text = feedsstatuscheck.feed_status_msgs[self.var]
-			else:
-				status_text = _('Feeds status: Unexpected')
-			self['feedStatusMSG'].setText(status_text)
-			self['tl_off'].hide()
-			self['tl_red'].hide()
-			self['tl_yellow'].hide()
-			self['tl_green'].hide()
-			if self.var == 'unstable':
-				self['tl_red'].show()
-			elif self.var == 'updating':
-				self['tl_yellow'].show()
-			elif self.var == 'stable':
-				self['tl_green'].show()
-			else:
-				self['tl_off'].show()
 
 	def autoResize(self):
 		desktop_w = enigma.getDesktop(0).size().width()
